backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
transformers==4.41.2
torch==2.3.1
torchaudio==2.3.1
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.3.0
